+++ tests/testthat/test-projection-forward.R
@@ -161,7 +161,7 @@ test_that("forward projection handles antimeridian", {
 
 test_that("forward projection matches reference data", {
   skip_on_cran()
-  hexify_build_icosa(11.25, 58.28252559, 0)
+  hexify_build_icosa(11.25, 58.282525588538995, 0)
   hexify_set_precision("ultra")
 
   truth <- read.csv(testthat::test_path("data/truth_txty.csv"),